# public/openapi.yaml
openapi: 3.1.0
info:
  title: DBackup API
  version: 3.4.0
  description: |
    REST API for DBackup - a self-hosted database backup automation platform with encryption, compression, and smart retention.

    ## Authentication

    DBackup supports two authentication methods:

    - **Session Authentication** - Used automatically when logged in via the web UI. Session cookies are sent with each request.
    - **API Key Authentication** - For scripts, CI/CD pipelines, and external integrations. Create an API key under **Access Management → API Keys**.

    > API keys do not inherit SuperAdmin privileges. Only explicitly assigned permissions are available.

    ## Rate Limits

    | Request Type | Default Limit |
    |---|---|
    | Authentication (login, etc.) | 5/min per IP |
    | GET requests | 100/min per IP |
    | POST / PUT / DELETE | 20/min per IP |

    Rate limits are configurable in Settings → Rate Limits.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
...
servers:
  - url: "{protocol}://{host}/api"
    description: DBackup Instance
    variables:
      protocol:
        default: https
        enum: [http, https]
      host:
        default: your-dbackup-instance.com
security:
  - BearerAuth: []
  - SessionAuth: []
...
paths:
...
  /jobs/bulk:
    post:
      tags: [Jobs]
      summary: Bulk Job Action
      description: |
        Applies one action to several jobs in a single request.

        A job that could not be processed does not fail the request. The response is always
        `200` with per-job outcomes in `data`, so a partial result is reported rather than
        discarded.

        **Required permission:** `jobs:write`
      operationId: bulkJobAction
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, ids]
              properties:
                action:
                  type: string
                  enum: [delete, enable, disable]
                ids:
                  type: array
                  items: { type: string }
                  minItems: 1
                  maxItems: 200
      responses:
        "200":
          description: Action applied, with per-job outcomes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /jobs/{id}:
    put:
      tags: [Jobs]
      summary: Update Job
      operationId: updateJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobRequest"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "409":
          description: Another job with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Jobs]
      summary: Delete Job
      operationId: deleteJob
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Job deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /adapters/{id}/clone:
    post:
      tags: [Adapters]
      summary: Clone Adapter
      description: |
        Creates a copy of the adapter configuration with the suffix `(Copy)` appended to
        the name. The same write permission as the adapter type is required:
        - `database` - `sources:write`
        - `storage` - `destinations:write`
        - `notification` - `notifications:write`
      operationId: cloneAdapter
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "201":
          description: Cloned adapter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdapterConfig"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API Key authentication. Create an API key under **Access Management → API Keys**.

        Format: `dbackup_xxxxxxxx`
    SessionAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie (browser login)
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource ID
  responses:
    Unauthorized:
      description: Missing, invalid, disabled, or expired credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Valid credentials but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Human-readable error message"
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    BulkResponse:
      type: object
      description: |
        Outcome of an action applied to several records.

        `success` describes the request, not the records. It is `true` whenever the request
        was handled, including when every record failed - the per-record verdicts live in
        `data`. Only authentication, permission, validation and unexpected server errors
        produce `success: false` with a 4xx or 5xx status.
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: One-line summary, for example "8 of 10 jobs deleted".
          example: 8 of 10 jobs deleted
        data:
          type: object
          properties:
            succeeded:
              type: array
              description: Ids (or paths, for storage) that were processed.
              items: { type: string }
            failed:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                    description: Human-readable label, when one is known.
                  error:
                    type: string
                    description: Why this record was not processed.
...
    Job:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Daily MySQL Backup"
        schedule:
          type: string
          example: "0 2 * * *"
        enabled:
          type: boolean
        sourceId:
          type: string
        compression:
          type: string
          enum: [NONE, GZIP, BROTLI]
        encryptionProfileId:
          type: [string, "null"]
        databases:
          type: string
          description: JSON array of database names
          example: '["mydb", "analytics"]'
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/JobDestination"
        source:
          type: object
          description: The source connection, never with its config, which holds hosts and secrets.
          properties:
            id:
              type: string
            name:
              type: string
            type:
              type: string
            adapterId:
              type: string
        createdAt:
          type: string
          format: date-time
    JobDestination:
      type: object
      properties:
        id:
          type: string
        configId:
          type: string
        priority:
          type: integer
        retention:
          type: string
          description: JSON retention policy
        config:
          type: object
          description: The destination connection, never with its config.
          properties:
            id:
              type: string
            name:
              type: string
            type:
              type: string
            adapterId:
              type: string
...
    UpdateJobRequest:
      type: object
      properties:
        name:
          type: string
        schedule:
          type: string
        sourceId:
          type: string
        destinations:
          type: array
          items:
            type: object
            required: [configId]
            properties:
              configId:
                type: string
              priority:
                type: integer
              retention:
                type: object
        enabled:
          type: boolean
        compression:
          type: string
          enum: [NONE, GZIP, BROTLI]
        encryptionProfileId:
          type: [string, "null"]
        notificationIds:
          type: array
          items:
            type: string
        notificationEvents:
          type: object
        databases:
          type: array
          items:
            type: string

    # ─── Executions ──────────────────────────────────────────────
...
    AdapterConfig:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          example: "Production MySQL"
        type:
          type: string
          enum: [database, storage, notification]
        adapterId:
          type: string
          description: Adapter type identifier
          example: mysql
        config:
          type: object
          description: Adapter-specific configuration (decrypted)
        metadata:
          type: [object, "null"]
          description: "Auto-detected metadata (version, status, etc.)"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        storageRole:
          type: string
          enum: [DESTINATION, SOURCE]
          default: DESTINATION
          description: >-
            Storage adapters only. A config either receives backups (DESTINATION) or
            provides folders to back up (SOURCE) - never both, because a destination writes
            job and chain folders into its configured path while a source reads folders out
            of it. Ignored for database and notification adapters.
...